add Quantizable class and division convenience functions
This commit adds some functionality to make division more convenient.

Most importantly, it adds a Quantizable class for rounding number
types to integers. This is used to implement floor, ceiling, and round
in a generic way.

Next, some convenience functions are added for supporting integer and
float division. These are common, monomorphic functions that make use
of the generic machinery.

Lastly, the introduction is improved to discuss division and how to
work with it.