rewrite unwrappable
to have a single operator, `unwrap-or-else`, which takes a `fail` continuation as an
argument; define `with-default`, `unwrap` and `expect` in terms of `unwrap-or-else`.

implementing `unwrap` and `expect` in classes.lisp requires moving `error` from
builtin.lisp to classes.lisp. i think that possibly `error` should be an actual builtin,
exported from the `coalton` package, but for now it remains in the standard library and is
re-exported from builtin.lisp, as it feels more at home there.