Respect `*print-escape*`; bind it when generating instance codegen syms
This PR rewrites all the various pprint-* functions on Coalton internal data structures to
recurse using write instead of format _ "~A". This way, printing those objects respects
the printer control variables like *print-escape*. Now, if you format a ty with ~S, you'll
see package names, whereas if you format one with ~A, you'll get the same output as
before.

The impetus for this change was my detecting a codegen bug where instance codegen syms
included the name but not the package of the types to which they applied, so (Into foo:Foo
bar:Foo) and (Into bar:Foo foo:Foo) got the same codegen symbol. Now, instance codegen
symbols are generated with *print-escape* bound to t, so the package names are included,
and there is no conflict.