test for substring appearance
Define two functions, `string:substring-index` and `string:substring?`, for searching a
string for the appearance of a substring.

`substring-index` is `String -> String -> (Optional UFix)`, returning the index of the
start of the substring.

`substring?` is `String -> String -> Boolean`, returning true if the substring appears in
the string.

Both of these functions take the "needle" substring as their first argument, and the
"haystack" superstring as their second argument. I believe this to be more useful for
currying; `(substring? "foo")` is a function which, given a string, tests if "foo" appears
in that string, and `(substring-index "foo")` is a function which, given a string, returns
the first index of "foo" in that string.