propagate constants among recursive let bindings
Previously, constant `let` bindings were propagated into the bodies
of the `let` expressions, but not into the remaining bindings.

For example, the following code would not compile because the binding
for `p` would be deleted after propagating into the body without having
propagated into the value of `q`.
```lisp
(let ((q (1+ p))
      (p 3))
  q)
```

This commit fixes that by propagating the values of all new constant
bindings into the values of the remaining nonconstant bindings.