Fix Fiasco error reporting from define-test
When a `(is (== ...))` test fails in `define-test`, Fiasco tries
to report the failure (unless it is running all the tests in batch).
However, that reporting routine in Fiasco triggers another error, yielding
a mysterious error message somethign like "The value NIL is not of type NUMBER",
which doesn't help to nail the failure at all.

It was because of the interaction of these three factors

(1) Fiasco reports failed tests with `describe-failed-tests` in suite.lisp,
    which calls `describe-object` on the conditions.
(2) Fiasco specializes `describe-object` method on several internal
    condition types.  What's relevant here is `failed-condition`, which
    is a subclass condition of `failure`, with customized reporter.
(3) Coalton defines `coalton-failure` subclassing `fiasco::failure`
    to represent failure during coalton tests.

The issue is that Fiasco does not specializes `describe-object` on
`failure`.  And SBCL does not like that generic function is called
on an object without a specialized method.

The user code is not supposed to call `describe-object` directly; it
should call `decribe` instead.  So it is partly Fiasco's problem.
Nonetheless, to make Fiasco reporting work, we should make
`coalton-failure` subclass `fiasco::failed-condition`.