Fix running tests in slime
In slime REPL, readtable is somewhat messed up once a complete test
is run, and (asdf:test-system :coalton/tests) fails in the second time.

This is a band-aid to ensure readtables are reset while it matters.
A proper fix may be done in slime, or there may be a workaround in
the Coalton side.  Meanwhile, some fix is nice since this is extemely
annoying while developing in slime.