Fix a bug with nested fundep instances
Previously, when one type would contain an instance of a class and it used that contained instance's instance to supply its own instance of the class, then it would fail to compile.

The issue was that define-instance removes the class constraints before
checking the types of the instance methods. Normally that's not a problem because they're entailed. But fundep-entail% checks for structural similarity, not entailment. To fix it, fundep-entail now skips fundep-entail% for any predicate that was entailed by the environment.

The tests pass, and the commit includes a regression test that failed with the fix removed and passes now.